backend/requirements.txt
# AI-Financial-Agent Python Dependencies

# FastAPI Framework
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Database
motor>=3.1.1
pymongo>=4.3.3

# M-Pesa Integration 
requests>=2.28.2
httpx>=0.24.0

# AI and Machine Learning
google-generativeai>=0.3.0

# Utilities
python-dateutil>=2.8.2
PyJWT>=2.6.0
python-multipart>=0.0.6
passlib>=1.7.4
bcrypt>=4.0.1
emails>=0.6.0
aiofiles>=23.1.0

# OCR and Image Processing
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
python-magic>=0.4.27

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0

#Biggie Requirements
mongoengine>=0.27.0
python-jose[cryptography]>=3.3.0
sendgrid>=6.10.0
mailersend>=0.5.0
africastalking>=1.2.5
celery[redis]>=5.3.4
redis>=5.0.1
reportlab>=4.0.7
pillow>=10.1.0
qrcode==7.4.2












